+++ apps/pleiepenger-i-livets-sluttfase-soknad/src/app/components/formik-file-uploader/FormikFileUploader.tsx
@@ -1,7 +1,12 @@
 import { Attachment } from '@navikt/sif-common-core-ds/lib/types/Attachment';
-import { VALID_EXTENSIONS } from '@navikt/sif-common-core-ds/lib/utils/attachmentUtils';
-import { FormikFileInput, TypedFormInputValidationProps, ValidationError } from '@navikt/sif-common-formik-ds/lib';
-import { ApiEndpoint } from '../../api/api';
+import {
+    FileDropAcceptImagesAndPdf,
+    TypedFormInputValidationProps,
+    ValidationError,
+} from '@navikt/sif-common-formik-ds/lib';
+import FormikFileDropInput from '@navikt/sif-common-formik-ds/lib/components/formik-file-drop-input/FormikFileDropInput';
+import api, { ApiEndpoint } from '../../api/api';
+import { getAttachmentURLFrontend } from '../../utils/attachmentUtilsAuthToken';
 import { useFormikFileUploader } from './useFormikFileUploader';
 
 interface Props extends TypedFormInputValidationProps<any, ValidationError> {
@@ -10,7 +15,7 @@ interface Props extends TypedFormInputValidationProps<any, ValidationError> {
     legend?: string;
     buttonLabel: string;
     apiEndpoint: ApiEndpoint;
-
+    onFilesUploaded?: (antall: number, antallFeilet: number) => void;
     onFileInputClick?: () => void;
     onErrorUploadingAttachments: (files: File[]) => void;
     onUnauthorizedOrForbiddenUpload: () => void;
@@ -22,23 +27,26 @@ function FormikFileUploader({
     legend,
     apiEndpoint,
     onFileInputClick,
+    onFilesUploaded,
     onErrorUploadingAttachments,
     onUnauthorizedOrForbiddenUpload,
 
     ...otherProps
 }: Props) {
     const { onFilesSelect } = useFormikFileUploader({
-        apiEndpoint,
+        value: attachments,
+        getAttachmentURLFrontend,
+        uploadFile: (file: File) => api.uploadFile(apiEndpoint, file),
         onUnauthorizedOrForbiddenUpload,
         onErrorUploadingAttachments,
-        value: attachments,
+        onFilesUploaded,
     });
 
     return (
-        <FormikFileInput
+        <FormikFileDropInput
             name={name}
             legend={legend || 'Dokumenter'}
-            accept={VALID_EXTENSIONS.join(', ')}
+            accept={FileDropAcceptImagesAndPdf}
             onFilesSelect={onFilesSelect}
             onClick={onFileInputClick}
             {...otherProps}